Make arrays V and absorp_coeff allocatable.
These arrays take a lot of memory, similarly to U (minus ghost zones), as they are 4D arrays.

Committing them to the static memory was leading to relocation truncated error R_X86_64_PC32 when running on a snapshot of size nz=1280,nyloc=60,nxloc=60,nw=25 as of Dec 2024.

Making these arrays allocatable solves the problem (even when the arrays are allocated).